Require dates to be formally correct
This format of regex will disallow clearly wrong dates, like 00
as date or something else than 01-12 as month etc. Additionally it
will require for the separator character to be correct for the
corresponding century, so + for 1800 etc.

I made the choice to only allow years up to 2029, so this would
require updating every ten years or so.